requirements.txt
psutil>=5.9.0
py-cpuinfo>=9.0.0
GPUtil>=1.4.0
torch>=2.0.0 